refactor: move throttling inside models
Avoid circular imports
